Make grep subcommand only call grep methods that exist.
Previously, we were using the list of modules on the first matching
matching and then calling grep on all of them. This broke on modules
that don't have a grep method.

Now we look for the list of methods on each machine, and only call
those modules that report a grep method.

We also changed the search order from being module wide (aka, check
rpms.grep across all machines, then foo.grep) to being machine wide
(check all grep methods on machine1, then machine2, etc). Not sure
this is better or worse, but it was easier.

Also change the output format a little bit. Also up for debate.