fixup puppet minions a bit
add a puppet_signed_certs_dir option to point to where puppet stores
the pem files. This is necessary in order to completely know about hosts
which were not revoked but were --cleaned from the puppetdb

also make sure we're looking for revoked serials only after sorting
for the latest serials per host.